Share the content index snapshot cache
