+++ src/Borea.Core/Index/IContentIndexSnapshotProvider.cs
@@ -1,0 +1,7 @@
+namespace Borea.Core.Index;
+
+/// <summary>Provides one shared, refreshed content index snapshot to index repositories.</summary>
+public interface IContentIndexSnapshotProvider
+{
+    Task<ContentIndexSnapshot> GetSnapshotAsync(CancellationToken cancellationToken = default);
+}
